removed unused routes
